@@ -10,24 +10,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.2.12",
-    "@angular/common": "^16.2.12",
-    "@angular/compiler": "^16.2.12",
-    "@angular/core": "^16.2.12",
-    "@angular/forms": "^16.2.12",
-    "@angular/platform-browser": "^16.2.12",
-    "@angular/platform-browser-dynamic": "^16.2.12",
-    "@angular/router": "^16.2.12",
+    "@angular/animations": "^17.3.3",
+    "@angular/common": "^17.3.3",
+    "@angular/compiler": "^17.3.3",
+    "@angular/core": "^17.3.3",
+    "@angular/forms": "^17.3.3",
+    "@angular/platform-browser": "^17.3.3",
+    "@angular/platform-browser-dynamic": "^17.3.3",
+    "@angular/router": "^17.3.3",
     "nanoid": "^4.0.1",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.13.3"
+    "zone.js": "~0.14.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.2.13",
-    "@angular/cli": "~16.2.13",
-    "@angular/compiler-cli": "^16.2.12",
-    "@tailwindcss/forms": "^0.5.3",
+    "@angular-devkit/build-angular": "^17.3.3",
+    "@angular/cli": "~17.3.3",
+    "@angular/compiler-cli": "^17.3.3",
+    "@tailwindcss/forms": "^0.5.7",
     "@types/jasmine": "~4.3.0",
     "autoprefixer": "^10.4.13",
     "jasmine-core": "~4.5.0",
@@ -38,7 +38,7 @@
     "karma-jasmine-html-reporter": "~2.0.0",
     "postcss": "^8.4.21",
     "prettier": "2.8.4",
-    "tailwindcss": "^3.2.7",
+    "tailwindcss": "^3.4.3",
     "typescript": "~5.4.4"
   }
 }
